clean and tests
